+++ changelog.md
@@ -5,6 +5,7 @@ Steam 'n' Rails 1.6.13 alpha
 Bug Fixes
 - Fix instanced carriage rendering after (de)coupling #690
 - Fix wide scotch yoke bogey model #686
+- Fix conductor toolbox crash #671
 ------------------------------------------------------
 Steam 'n' Rails 1.6.12 alpha
 ------------------------------------------------------